Move new functions for processed fba and fbs data from common to its own module for cleaner organization
